build: clear compiler warnings flagged by CI
- vt_ops.c: consume write() return in vt_reply with a short-write loop
  (retry on EINTR), silencing the musl warn_unused_result warning
- bd_verb.c: size the parse scratch buffer to the caller token size so
  snprintf into the 128-byte class/chain dests no longer looks truncating
- bd_log.c: clamp struct tm fields and milliseconds to valid ranges,
  removing the -Wformat-truncation warnings and hardening timestamps
- bd_mxp.c: cast the guarded positive rune value to unsigned for the
  BD_UTF8_RUNE_MAX comparison, fixing the arm32 -Wsign-compare warning

+++ src/birdie/bd_mxp.c
@@ -173,7 +173,7 @@ finish_entity(bd_mxp *m)
 			v = strtol(e + 1, &end, 10);
 		unsigned char u[BD_UTF8_MAX];
 		int ul;
-		if (*end == '\0' && v > 0 && v <= BD_UTF8_RUNE_MAX &&
+		if (*end == '\0' && v > 0 && (unsigned long)v <= BD_UTF8_RUNE_MAX &&
 		    (ul = bd_utf8_encode(u, (uint32_t)v)) > 0) {
 			out_put(m, u, (size_t)ul);   /* handles all planes */
 		} else {                        /* malformed/invalid: verbatim */
